+++ packages/suite/src/actions/suite/storageActions.ts
@@ -352,6 +352,15 @@ export const saveMessageSystem = () => async (_dispatch: Dispatch, getState: Get
     );
 };
 
+export const saveFirmware = () => async (_dispatch: Dispatch, getState: GetState) => {
+    console.log('storageActions.saveFirmware');
+    if (!(await isDBAccessible())) return;
+    const { firmware } = getState();
+    console.log('storageActions.saveFirmware 2', firmware);
+
+    db.addItem('firmware', { firmwareHashInvalid: firmware.firmwareHashInvalid }, 'firmware', true);
+};
+
 export const removeDatabase = () => async (dispatch: Dispatch, getState: GetState) => {
     if (!(await isDBAccessible())) return;
 
@@ -400,6 +409,7 @@ export const loadStorage = () => async (dispatch: Dispatch, getState: GetState) 
         const mappedTxs: AppState['wallet']['transactions']['transactions'] = {};
         const messageSystem = await db.getItemByPK('messageSystem', 'suite');
         const backendSettings = await db.getItemsWithKeys('backendSettings');
+        const firmware = await db.getItemByPK('firmware', 'firmware');
 
         txs.forEach(item => {
             const k = getAccountKey(item.tx.descriptor, item.tx.symbol, item.tx.deviceState);
@@ -495,6 +505,10 @@ export const loadStorage = () => async (dispatch: Dispatch, getState: GetState) 
                 messageSystem: {
                     ...initialState.messageSystem,
                     ...messageSystem,
+                },
+                firmware: {
+                    ...initialState.firmware,
+                    ...firmware,
                 },
             },
         });